+++ web/tests/EducationBenchmarking.Web.E2ETests/Features/CompareYourCosts.feature
@@ -1,4 +1,5 @@
-Feature: SchoolHomePage
+Feature: CompareYourCosts
+
 	School HomePage is showing correct data
 
 Feature: Expenditure Benchmarking for Schools
