+++ web/tests/EducationBenchmarking.Web.E2ETests/Pages/CompareYourCostsPage.cs
@@ -1,0 +1,6 @@
+namespace EducationBenchmarking.Web.E2ETests.Pages;
+
+public class CompareYourCostsPage
+{
+    
+}
