+++ web/tests/EducationBenchmarking.Web.E2ETests/Pages/SearchPage.cs
@@ -1,0 +1,6 @@
+namespace EducationBenchmarking.Web.E2ETests.Pages;
+
+public class SearchPage
+{
+    
+}
